Extract sample_url_request_callback.dart & sample-executor.dart
