+++ lib/cronet_flutter.dart
@@ -10,6 +10,9 @@ import 'src/cronet.dart';
 
 export 'src/cronet.dart';
 
+part 'src/sample_executor.dart';
+part 'src/sample_url_request_callback.dart';
+
 final DynamicLibrary Function() loadLibrary = () {
   if (Platform.isWindows) {
     return DynamicLibrary.open('cronet.86.0.4240.198.dll');
@@ -180,95 +183,3 @@ class CronetFlutter {
     }
   }
 }
-
-// --------------------------------------------------------------------------------------------------------------------------------
-
-class SampleUrlRequestCallback extends Opaque {}
-
-typedef _c_SampleUrlRequestCallback_Create = Pointer<SampleUrlRequestCallback>
-    Function();
-
-typedef _dart_SampleUrlRequestCallback_Create
-    = Pointer<SampleUrlRequestCallback> Function();
-
-final _SampleUrlRequestCallback_Create_ptr =
-    nativeInteropLib.lookup<NativeFunction<_c_SampleUrlRequestCallback_Create>>(
-        'SampleUrlRequestCallback_Create');
-final _dart_SampleUrlRequestCallback_Create _SampleUrlRequestCallback_Create =
-    _SampleUrlRequestCallback_Create_ptr.asFunction<
-        _dart_SampleUrlRequestCallback_Create>();
-
-// --------------------------------
-
-typedef _c_SampleUrlRequestCallback_Destory = Void Function(
-    Pointer<SampleUrlRequestCallback> request_callback);
-
-typedef _dart_SampleUrlRequestCallback_Destory = void Function(
-    Pointer<SampleUrlRequestCallback> request_callback);
-
-final _SampleUrlRequestCallback_Destory_ptr = nativeInteropLib
-    .lookup<NativeFunction<_c_SampleUrlRequestCallback_Destory>>(
-        'SampleUrlRequestCallback_Destory');
-final _dart_SampleUrlRequestCallback_Destory _SampleUrlRequestCallback_Destory =
-    _SampleUrlRequestCallback_Destory_ptr.asFunction<
-        _dart_SampleUrlRequestCallback_Destory>();
-
-// --------------------------------
-
-typedef _c_SampleUrlRequestCallback_GetUrlRequestCallback
-    = Pointer<Cronet_UrlRequestCallback> Function(
-        Pointer<SampleUrlRequestCallback> request_callback);
-
-typedef _dart_SampleUrlRequestCallback_GetUrlRequestCallback
-    = Pointer<Cronet_UrlRequestCallback> Function(
-        Pointer<SampleUrlRequestCallback> request_callback);
-
-final _SampleUrlRequestCallback_GetUrlRequestCallback_ptr = nativeInteropLib
-    .lookup<NativeFunction<_c_SampleUrlRequestCallback_GetUrlRequestCallback>>(
-        'SampleUrlRequestCallback_GetUrlRequestCallback');
-final _dart_SampleUrlRequestCallback_GetUrlRequestCallback
-    _SampleUrlRequestCallback_GetUrlRequestCallback =
-    _SampleUrlRequestCallback_GetUrlRequestCallback_ptr.asFunction<
-        _dart_SampleUrlRequestCallback_GetUrlRequestCallback>();
-
-// --------------------------------------------------------------------------------------------------------------------------------
-
-class SampleExecutor extends Opaque {}
-
-typedef _c_SampleExecutor_Create = Pointer<SampleExecutor> Function();
-
-typedef _dart_SampleExecutor_Create = Pointer<SampleExecutor> Function();
-
-final _SampleExecutor_Create_ptr = nativeInteropLib
-    .lookup<NativeFunction<_c_SampleExecutor_Create>>('SampleExecutor_Create');
-final _dart_SampleExecutor_Create _SampleExecutor_Create =
-    _SampleExecutor_Create_ptr.asFunction<_dart_SampleExecutor_Create>();
-
-// --------------------------------
-
-typedef _c_SampleExecutor_Destory = Void Function(
-    Pointer<SampleExecutor> executor);
-
-typedef _dart_SampleExecutor_Destory = void Function(
-    Pointer<SampleExecutor> executor);
-
-final _SampleExecutor_Destory_ptr =
-    nativeInteropLib.lookup<NativeFunction<_c_SampleExecutor_Destory>>(
-        'SampleExecutor_Destory');
-final _dart_SampleExecutor_Destory _SampleExecutor_Destory =
-    _SampleExecutor_Destory_ptr.asFunction<_dart_SampleExecutor_Destory>();
-
-// --------------------------------
-
-typedef _c_SampleExecutor_GetExecutor = Pointer<Cronet_Executor> Function(
-    Pointer<SampleExecutor> executor);
-
-typedef _dart_SampleExecutor_GetExecutor = Pointer<Cronet_Executor> Function(
-    Pointer<SampleExecutor> executor);
-
-final _SampleExecutor_GetExecutor_ptr =
-    nativeInteropLib.lookup<NativeFunction<_c_SampleExecutor_GetExecutor>>(
-        'SampleExecutor_GetExecutor');
-final _dart_SampleExecutor_GetExecutor _SampleExecutor_GetExecutor =
-    _SampleExecutor_GetExecutor_ptr.asFunction<
-        _dart_SampleExecutor_GetExecutor>();
